refactor(pragma): move error module to folder structure per code standards
Split error.ts into error/ folder: constants.ts (ErrorCode enum),
PragmaError.ts (single class export), PragmaErrorData.ts (single
interface export), barrel index.ts. Follows code/export/shape and
code/naming/single-export-file standards.
