+++ packages/cli/pragma/src/error/index.ts
@@ -1,0 +1,4 @@
+export type { ErrorCode } from "./constants.js";
+export { ERROR_CODES } from "./constants.js";
+export { PragmaError } from "./PragmaError.js";
+export type { PragmaErrorData } from "./PragmaErrorData.js";
