+++ packages/cli/pragma/src/index.ts
@@ -2,7 +2,7 @@ export type { PragmaConfig } from "./config.js";
 export { readConfig } from "./config.js";
 export type { Channel } from "./constants.js";
 export { VALID_CHANNELS } from "./constants.js";
-export type { ErrorCode, PragmaErrorData } from "./error.js";
-export { ERROR_CODES, PragmaError } from "./error.js";
+export type { ErrorCode, PragmaErrorData } from "./error/index.js";
+export { ERROR_CODES, PragmaError } from "./error/index.js";
 export type { PackageManager } from "./pm.js";
 export { detectLocalInstall, detectPackageManager, PM_COMMANDS } from "./pm.js";
